refactor(schools): extract filter configurations and distance calculation logic
- Create composables/useSchoolDistance.ts for reactive distance calculation
  - Encapsulates distance cache computation from home location to all schools
  - Accepts schools and homeLocation as ComputedRef parameters
  - Returns distanceCache computed ref for reactive updates

- Create utils/schoolFilterConfigs.ts factory for filter configurations
  - Encapsulates all 7 filter configs (name, division, status, favorite, state, fit_score, distance)
  - Maintains complex filter logic including state extraction and distance filtering
  - Properly typed with FilterConfig and FilterValue types

- Refactor pages/schools/index.vue
  - Remove inline distanceCache computed (135 lines)
  - Remove inline filterConfigs array (103 lines)
  - Import and use useSchoolDistance composable
  - Import and use createSchoolFilterConfigs factory
  - Improves maintainability and enables reusability of distance calculation logic

Verification:
- TypeScript type checking: PASS
- All 4994 unit tests: PASS
- No breaking changes to existing functionality
